refactor: getIsGuest 를 auth 유틸로 통합
- 토큰 기반 role 판정(getRoleFromToken·getIsGuest)을 auth.ts 한 곳으로 모음 — 소비처가 전부 RSC 라 next/headers 유입 문제 없음
- server-only 마커 패키지 추가도 검토했으나 next/headers import 가 이미 클라이언트 유입을 빌드 에러로 차단해 패키지 없이 NOTE 주석만 유지
